Send a proper error if an activity doesn't exist
Before this commit, server used to return an internal server error if an
activity didn't exist. This is fixed rather easily, simply by replacing
`.fetch_one` call with `.fetch_optional` call, which returns
`Option<PgRow>`. After that, the code checks if that `Option` is `None`,
if it is, returns `Err(ApiError::ResourceNotFound)`, it it isn't, it
simply unwraps the `Option`.

Signed-off-by: asyncth <asyncthry@gmail.com>