Combine the check if a user exists and data fetch
Previously server executed an SQL query to find out if a user exists or
not before executing the main SQL query, which does the actual data
retrieval. This commit fixes this using the same method as the one used
in the previous commit: `.fetch_one` call is replaced with a call to
`.fetch_optional`, which returns an `Option<PgRow>`.

Signed-off-by: asyncth <asyncthry@gmail.com>