Migrate all collections to timestamp-based IDs
This commit changes the way pagination of inbox and outbox collections
works. Now it uses timestamp-based IDs instead of serial-based IDs,
which brings it closer to the way pagination works for followers and
following collections. Additionally, this commit fixes the bug that
caused the server to send `prev` field on the first page of collections.

Signed-off-by: asyncth <asyncthry@gmail.com>