Refactor collections code and add endpoints module
This commit completely refactors collections code by removing
copy-pasted code and introducing new types. First things first, this
commit introduces `activitypub::collections::Provider` trait, which
defines a few functions for fetching various pages. Implementations of
this trait contain collection-specific code to fetch data from the
database and present it as a `activitypub::collections::Items` enum,
which is really just a wrapper type around a vector of either of strings
(needed for collections in which ActivityStreams IDs are sent instead of
ActivityStreams objects themselves) or `BaseBox`es.

To abstract away majority of the boilerplate code, this commit adds
`activitypub::collections::Collection` struct, which requires an
implementation of the `Provider` trait to function. This struct contains
code that isn't specific to a particular collection and is shared
between all collections, such as code that figures out the link to the
next and the previous pages, creating ActivityStreams collection
objects, setting correct values for each field, etc.

This commit also introduces a new top-level `endpoints` module, which
from now on will contain all HTTP API endpoints. New versions of
collection endpoints that use `Collection` struct and `Provider` trait
are already placed into this new module.

Signed-off-by: asyncth <asyncthry@gmail.com>