Keep public addressing in object sanitizer
Fixes an issue that caused object sanitizer to potentially remove a
public addressing URI. To be exact, `.take(5)` call took 5 first
elements of `cc` and `to` arrays and discarded the rest. If a client
sends 10 elements in `to` array and a public addressing URI as, for
example, 8th element, it would be removed by the server and the post
would become private, which is undesirable.

Signed-off-by: asyncth <asyncthry@gmail.com>