Improve mentioning followers of an actor
This commit accomplishes several things. First of all, the server now
merges values of `to` and `cc` fields on the object and on the activity
if a `Create` activity is submitted to the outbox. Next, server will now
recognize if the submitted object/activity doesn't have a public
addressing URI in `to` or `cc` fields and the server will mark the
activity as non-public in the database, which means that only mentioned
actors will be able to see it. Finally, this commit fixes the problem
that is mentioned in the commit before the previous commit by
introducing two new columns: `to_followers_of` and `cc_followers_of`.
Those two columns store IDs of users, who's followers are mentioned by
the object/activity.

BREAKING CHANGE: this commit changes the database schema and provides no
migration for it. If you host an instance, you must remove all tables
and start from scratch.

Signed-off-by: asyncth <asyncthry@gmail.com>