Retry only deliveries that failed with a net error
So far every failed delivery was put into a retry queue. ActivityPub
spec, however, says that deliveries should be retried if they failed
with a network error. This commit places deliveries into the retry queue
only if the future responsible for sending the request and receiving the
status code outputs an error.

Signed-off-by: asyncth <asyncthry@gmail.com>