Implement retrying failed deliveries
This adds an async task that runs alongside the HTTP server, it picks a
failed delivery from the queue, waits until an hour has passed since it
was tried and retries it. If this fails as well, it gets put back into
the queue. Failed deliveries are retried for a maximum of three times.

Signed-off-by: asyncth <asyncthry@gmail.com>